Pull up POM load and processing logic.
